add go-sdk and update related examples
