+++ avail-go/examples/go-examples/config.json
