+++ avail-go/go.mod
@@ -1,0 +1,30 @@
+module avail-go-sdk
+
+go 1.21.0
+
+require (
+	github.com/centrifuge/go-substrate-rpc-client/v4 v4.2.1
+	github.com/vedhavyas/go-subkey v1.0.4
+)
+
+require (
+	github.com/ChainSafe/go-schnorrkel v1.0.0 // indirect
+	github.com/cosmos/go-bip39 v1.0.0 // indirect
+	github.com/deckarep/golang-set v1.8.0 // indirect
+	github.com/decred/base58 v1.0.4 // indirect
+	github.com/decred/dcrd/crypto/blake256 v1.0.0 // indirect
+	github.com/ethereum/go-ethereum v1.10.20 // indirect
+	github.com/go-stack/stack v1.8.1 // indirect
+	github.com/gorilla/websocket v1.5.0 // indirect
+	github.com/gtank/merlin v0.1.1 // indirect
+	github.com/gtank/ristretto255 v0.1.2 // indirect
+	github.com/mimoo/StrobeGo v0.0.0-20220103164710-9a04d6ca976b // indirect
+	github.com/pierrec/xxHash v0.1.5 // indirect
+	github.com/rs/cors v1.8.2 // indirect
+	github.com/vedhavyas/go-subkey/v2 v2.0.0 // indirect
+	golang.org/x/crypto v0.7.0 // indirect
+	golang.org/x/sys v0.6.0 // indirect
+	gopkg.in/natefinch/npipe.v2 v2.0.0-20160621034901-c1b8fa8bdcce // indirect
+)
+
+// replace github.com/centrifuge/go-substrate-rpc-client/v4 => /Users/kailaskr/gsrpc
