+++ avail-go/sdk/sdk.go
@@ -1,0 +1,32 @@
+package sdk
+
+import (
+	"github.com/centrifuge/go-substrate-rpc-client/v4/client"
+	"github.com/centrifuge/go-substrate-rpc-client/v4/rpc"
+)
+
+type SubstrateAPI struct {
+	RPC    *rpc.RPC
+	Client client.Client
+}
+
+func NewSDK(url string) (*SubstrateAPI, error) {
+	cl, err := client.Connect(url)
+	if err != nil {
+		return nil, err
+	}
+
+	newRPC, err := rpc.NewRPC(cl)
+	if err != nil {
+		return nil, err
+	}
+
+	return &SubstrateAPI{
+		RPC:    newRPC,
+		Client: cl,
+	}, nil
+}
+
+func (api *SubstrateAPI) Close() {
+	api.Client.Close()
+}
