--- examples/go/internal/extrinsics/extrinsics.go
@@ -1,106 +1,0 @@
-package extrinsics
-
-import (
-	"crypto/rand"
-	"encoding/hex"
-	"fmt"
-	"time"
-
-	gsrpc "github.com/centrifuge/go-substrate-rpc-client/v4"
-	"github.com/centrifuge/go-substrate-rpc-client/v4/signature"
-	"github.com/centrifuge/go-substrate-rpc-client/v4/types"
-)
-
-// SubmitData sends the extrinsic data to Substrate
-func SubmitData(api *gsrpc.SubstrateAPI, data string, seed string, appID int, finalizedBlockCh chan types.Hash) error {
-
-	meta, err := api.RPC.State.GetMetadataLatest()
-	if err != nil {
-		return err
-	}
-
-	c, err := types.NewCall(meta, "DataAvailability.submit_data", types.NewBytes([]byte(data)))
-	if err != nil {
-		return fmt.Errorf("error creating new call: %s", err)
-	}
-
-	// Create the extrinsic
-	ext := types.NewExtrinsic(c)
-
-	genesisHash, err := api.RPC.Chain.GetBlockHash(0)
-	if err != nil {
-		return fmt.Errorf("error getting genesis hash: %s", err)
-	}
-
-	rv, err := api.RPC.State.GetRuntimeVersionLatest()
-	if err != nil {
-		return fmt.Errorf("error retrieveing runtime version: %s", err)
-	}
-
-	keyringPair, err := signature.KeyringPairFromSecret(seed, 42)
-	if err != nil {
-		return fmt.Errorf("error creating keyring pair: %s", err)
-	}
-
-	key, err := types.CreateStorageKey(meta, "System", "Account", keyringPair.PublicKey)
-	if err != nil {
-		return fmt.Errorf("cannot create storage key:%w", err)
-	}
-
-	var accountInfo types.AccountInfo
-	ok, err := api.RPC.State.GetStorageLatest(key, &accountInfo)
-	if err != nil || !ok {
-		return fmt.Errorf("cannot get latest storage:%v", err)
-	}
-
-	nonce := uint32(accountInfo.Nonce)
-	options := types.SignatureOptions{
-		BlockHash:          genesisHash,
-		Era:                types.ExtrinsicEra{IsMortalEra: false},
-		GenesisHash:        genesisHash,
-		Nonce:              types.NewUCompactFromUInt(uint64(nonce)),
-		SpecVersion:        rv.SpecVersion,
-		Tip:                types.NewUCompactFromUInt(100),
-		AppID:              types.NewUCompactFromUInt(uint64(appID)),
-		TransactionVersion: rv.TransactionVersion,
-	}
-
-	// Sign the transaction using Alice's default account
-	err = ext.Sign(keyringPair, options)
-	if err != nil {
-		return fmt.Errorf("cannot sign:%v", err)
-	}
-
-	// Send the extrinsic
-	sub, err := api.RPC.Author.SubmitAndWatchExtrinsic(ext)
-	if err != nil {
-		return fmt.Errorf("cannot submit extrinsic:%v", err)
-	}
-
-	defer sub.Unsubscribe()
-	timeout := time.After(100 * time.Second)
-	for {
-		select {
-		case status := <-sub.Chan():
-			if status.IsInBlock {
-				fmt.Printf("Txn inside block %v\n", status.AsInBlock.Hex())
-			} else if status.IsFinalized {
-				fmt.Printf("Txn inside finalized block\n")
-				finalizedBlockCh <- status.AsFinalized
-				return nil
-			}
-		case <-timeout:
-			fmt.Printf("timeout of 100 seconds reached without getting finalized status for extrinsic")
-			return nil
-		}
-	}
-}
-
-// RandToken generates a random hex value.
-func RandToken(n int) (string, error) {
-	bytes := make([]byte, n)
-	if _, err := rand.Read(bytes); err != nil {
-		return "", err
-	}
-	return hex.EncodeToString(bytes), nil
-}
